refactor: Rename `session-created` to `connect`
