+++ frontend/src/app/sessions/user-sessions-wrapper/active-sessions/connect/guacamole/guacamole.component.css
